push

+++ c/typecast.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int main(void) {
+float main(void) {
   int i = 1000;
   printf("%d\n", (char)i);
 }
